+++ src/apis/config/privateApi.ts
@@ -27,6 +27,13 @@ privateApi.interceptors.request.use(
     return config;
   },
   (error) => {
+    alert('오류가 발생했습니다. 다시 시도해주세요!');
+    location.href = '/auth/logout';
+    // if (error.response.status === 415) {
+    //   if (confirm('세션이 만료되었습니다. 다시 로그인해주세요!')) {
+    //     open('/menu');
+    //   }
+    // }
     return Promise.reject(error);
   }
 );
