+++ src/app/(Sub)/menu/MenuListSection.tsx
@@ -1,10 +1,8 @@
 'use client';
 
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 
-import { PrimaryButton } from '@/components/Button';
-import { Popup } from '@/components/Modal';
+import LogoutModal from '@/app/(Main)/mypose/components/LogoutModal';
 import { useOverlay } from '@/components/Overlay/useOverlay';
 import { menuList } from '@/constants/data';
 import useUserState from '@/context/userState';
@@ -12,24 +10,9 @@ import useUserState from '@/context/userState';
 export default function MenuListSection() {
   const { isLogin } = useUserState();
   const { open } = useOverlay();
-  const router = useRouter();
 
   function handleLogout() {
-    open(({ exit }) => (
-      <Popup
-        title="로그아웃"
-        content={`북마크는 로그인 시에만 유지되어요.\n정말 로그아웃하시겠어요?`}
-      >
-        <>
-          <PrimaryButton
-            text={'로그아웃'}
-            type="secondary"
-            onClick={() => router.replace('/auth/logout')}
-          />
-          <PrimaryButton text="로그인 유지" onClick={exit} />
-        </>
-      </Popup>
-    ));
+    open(({ exit }) => <LogoutModal exit={exit} />);
   }
 
   return (
